Move crate error to error.rs
