+++ sim7000-async/src/error.rs
@@ -1,0 +1,25 @@
+use embassy_time::TimeoutError;
+
+use crate::at_command::SimError;
+
+#[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub enum Error {
+    InvalidUtf8,
+    BufferOverflow,
+    Sim(SimError),
+    Timeout,
+    Serial,
+}
+
+impl embedded_io::Error for Error {
+    fn kind(&self) -> embedded_io::ErrorKind {
+        embedded_io::ErrorKind::Other
+    }
+}
+
+impl From<TimeoutError> for Error {
+    fn from(_: TimeoutError) -> Self {
+        Error::Timeout
+    }
+}
